refactor: 클래스 object로 변경
