+++ src/main/kotlin/racingcar/Application.kt
@@ -1,14 +1,8 @@
 package racingcar
 
 import racingcar.controller.CarController
-import racingcar.view.InputView
-import racingcar.view.OutputView
 
 fun main() {
-    val carController = CarController(
-        inputView = InputView(),
-        outputView = OutputView(),
-        inputValidator = InputValidator()
-    )
+    val carController = CarController()
     carController.run()
 }
