+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -4,7 +4,7 @@ import racingcar.util.Message
 import java.io.BufferedReader
 import java.io.InputStreamReader
 
-class InputView {
+object InputView {
     fun inputCarNames(): List<String> {
         println(Message.REGISTER_CAR_NAMES)
         return br.readLine().split(INPUT_SEPARATOR)
@@ -15,8 +15,6 @@ class InputView {
         return br.readLine()
     }
 
-    companion object {
-        val br = BufferedReader(InputStreamReader(System.`in`))
-        const val INPUT_SEPARATOR = ","
-    }
+    private val br = BufferedReader(InputStreamReader(System.`in`))
+    private const val INPUT_SEPARATOR = ","
 }
